Move address-formatter back to its original package
