+++ libs/address-formatter/src/main/java/org/woheller69/AndroidAddressFormatter/RegexPatternCache.java
@@ -1,4 +1,4 @@
-package org.woheller69;
+package org.woheller69.AndroidAddressFormatter;
 
 import java.util.Map;
 import java.util.HashMap;
